requirements.txt
annotated-types==0.7.0
anyio==3.7.1
black==23.11.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorlog==6.8.0
distro==1.9.0
flake8==6.1.0
Flask==3.0.0
Flask-Cors==4.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.11
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
openai==1.3.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
psutil==5.9.6
pycodestyle==2.11.1
pydantic==2.12.2
pydantic_core==2.41.4
pyflakes==3.1.0
pytest==7.4.3
pytest-flask==1.3.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
Werkzeug==3.1.3
